backend/requirements/base.txt
# Core Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Celery (Task Queue)
celery==5.3.4

# Validation and Serialization
marshmallow==3.20.1
python-dotenv==1.0.0
pydantic==2.5.2

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Security
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.8.0

# Monitoring and Logging
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# WebSocket
python-socketio==5.10.0
eventlet==0.33.3

# Date/Time
arrow==1.3.0
